refactor(binary_io): remove it in favor of binary_io_new
